reorganized interface directory and sub-directories
