+++ actions/report.py
@@ -1,5 +1,7 @@
-def build_facility_report(arboretum):
-    for river in arboretum.rivers:
+from arboretum import Arboretum
+
+def build_facility_report(Arboretum):
+    for river in Arboretum.rivers:
         print(f'River [{river.id}]')
 
     input("\n\nPress any key to continue...")
