+++ interfaces/__init__.py
@@ -1,9 +1,9 @@
-from .terrestrial import ITerrestrial
+from interfaces.habitat.terrestrial import ITerrestrial
 from .animal.freshwater import IFreshwater
-from .aquatic import IAquatic
-from .flying import IFlying
-from .walking import IWalking
-from .swimming import ISwimming
+from interfaces.habitat.aquatic import IAquatic
+from .animal.flying import IFlying
+from .animal.walking import IWalking
+from .animal.swimming import ISwimming
 from .identifiable import Identifiable
-from .habitat.contains_animals import IContainsAnimals
-from .habitat.contains_plants import IContainsPlants
+from interfaces.habitat.contains_animals import IContainsAnimals
+from interfaces.habitat.contains_plants import IContainsPlants
