+++ interfaces/animal/__init__.py
@@ -1,1 +1,5 @@
 from .freshwater import IFreshwater
+from .flying import IFlying
+from .saltwater import ISaltwater
+from .swimming import ISwimming
+from .walking import IWalking
