+++ interfaces/animal/walking.py
